Check for previous zshrc

+++ oh-my-zsh/install.sh
@@ -2,4 +2,6 @@ sudo apt install -y zsh curl
 
 sh -c "$(curl -fsSL https://raw.githubusercontent.com/robbyrussell/oh-my-zsh/master/tools/install.sh | sed 's:env zsh -l::g' | sed 's:chsh -s .*$::g')"
 
-mv ~/.zshrc.pre-oh-my-zsh ~/.zshrc
+if [ -e ~/.zshrc.pre-oh-my-zsh ]; then
+	mv ~/.zshrc.pre-oh-my-zsh ~/.zshrc
+fi
